Support rectangular screen shapes
Previously we'd build the canvas from a square and make the box size
depend on the canvas shapes. However, this prohibits square playing
fields.

Fix this by instead making boxSize configurable and using that to build
a grid of those boxes, which makes the the canvas depend on the box size
instead of the other way around.