Implement GameMap and modernize Screen
These two changes went uncomitted a bit too long, so it'll have to be done together. Ditch the WallParser class which felt very clunky in favor of a GameMap that actually properly implements a map to play on, with full Json decoding and encoding support.

Screen now doesn't influence the DOM directly, but only the CanvasElement it gets passed.